fix: fix generating examples #42
